Reusing NodeDiskManagerConfig
Redeclaration of ConfigMap struct is removed. ConfigMap is now declared as an alias of NodeDiskManagerConfig
