+++ integration_test/ndm_util/path_filter_util.go
@@ -1,12 +1,18 @@
 package ndmutil
 
 import (
+	"fmt"
 	"github.com/ghodss/yaml"
 	sysutil "github.com/openebs/CITF/utils/system"
+	"github.com/openebs/node-disk-manager/cmd/controller"
 	"io/ioutil"
 	"strings"
 )
 
+type ConfigMap controller.NodeDiskManagerConfig
+
+// Replace the ConfigMap portion in the configuration
+// yaml with the given ConfigMap struct and apply it.
 func ReplaceAndApplyConfig(configMap ConfigMap) error {
 	yamlBytes, err := ioutil.ReadFile(GetNDMOperatorFilePath())
 	if err != nil {
@@ -20,4 +26,55 @@ func ReplaceAndApplyConfig(configMap ConfigMap) error {
 	s2 := strings.SplitN(yamlString, "---", 2)[1]
 	yamlString = s1 + "\n  node-disk-manager.config: | \n    " + stringConfig + "\n--- \n" + s2
 	return sysutil.RunCommandWithGivenStdin("kubectl apply -f -", yamlString)
+}
+
+// Get the ConfigMap from the NDMConfiguration file
+// ConfigMap is generated by parsing the configuration yaml
+func GetNDMConfig(fileName string) ConfigMap {
+	var configMap ConfigMap
+	data, err := ioutil.ReadFile(fileName)
+	if err != nil {
+		fmt.Println(err)
+	}
+	err = yaml.Unmarshal(data, &configMap)
+	if err != nil {
+		fmt.Println(err)
+	}
+	yamlString := string(data)
+	yamlString = strings.Split(yamlString, "---")[0]
+	yamlString = strings.Split(yamlString, "node-disk-manager.config: |")[1]
+	yaml.Unmarshal([]byte(yamlString), &configMap)
+	return configMap
+}
+
+// SetIncludePath is used to set the include section in
+// path filter of NDMConfig
+func (c *ConfigMap) SetIncludePath(deviceList ...string) {
+	for index, element := range c.FilterConfigs {
+		if element.Key == "path-filter" {
+			c.FilterConfigs[index].Include = strings.Join(deviceList, ",")
+			c.FilterConfigs[index].Exclude = ""
+		}
+	}
+}
+
+// SetExcludePath is used to set the exclude section in
+// path filter of NDMConfig
+func (c *ConfigMap) SetExcludePath(deviceList ...string) {
+	for index, element := range c.FilterConfigs {
+		if element.Key == "path-filter" {
+			c.FilterConfigs[index].Exclude = strings.Join(deviceList, ",")
+			c.FilterConfigs[index].Include = ""
+		}
+	}
+}
+
+// SetPathFilter is used to change the state of
+// path filter in NDMConfig
+func (c *ConfigMap) SetPathFilter(state string) {
+	for index, element := range c.FilterConfigs {
+		if element.Key == "path-filter" {
+			c.FilterConfigs[index].State = state
+		}
+	}
 }
